Create common dupSafeName function
