+++ apps/dashboard/app/javascript/packs/files/uppy_ops.js
@@ -3,6 +3,7 @@ import Dashboard from '@uppy/dashboard'
 import XHRUpload from '@uppy/xhr-upload'
 import _ from 'lodash';
 import {CONTENTID, EVENTNAME as DATATABLE_EVENTNAME} from './data_table.js';
+import { dupSafeName } from './utils.js';
 
 let uppy = null;
 
@@ -166,34 +167,6 @@ function getEmptyDirs(entry){
       })
     }
   });
-}
-
-function dupSafeName(originalName) {
-  const currentFilenames = history.state.currentFilenames;
-  const extIndex = originalName.lastIndexOf('.');
-  let newName, extension;
-  if (extIndex == -1) {
-    // If no extension or directory, disregard extension
-    newName = originalName;
-    extension = '';
-  } else {
-    newName = originalName.slice(0, extIndex);
-    extension = originalName.slice(extIndex);
-  }
-  // If originalName in cur dir, try `${originalName}_copy`.
-  if (currentFilenames.includes(newName + extension)) {
-    newName += '_copy';
-    // If `${originalName}_copy` exists, try `${originalName}_copy_{i}' starting at i=1 until a file doesn't exist
-    if (currentFilenames.includes(newName + extension)) {
-      let copyNumber = 1;
-      newName += `_${copyNumber}`;
-      while (currentFilenames.includes(newName + extension)) {
-        copyNumber++;
-        newName = newName.slice(0, newName.lastIndexOf('_') + 1) + copyNumber;
-      }
-    }
-  }
-  return `${newName}${extension}`
 }
 
 function renameIfDuplicate(currentFile, files) {
